Similar change with less user-facing differences
Instead of using a textarea, we make the div containing the citation
select on click. Also, fixed some issues with including two copies
of jquery and bootstrap, which were causing the modal to fire twice.

This commit also introduces a `main.js` file for keeping together
miscellaneous JS that is used on multiple pages.