Simplify .travis.yml
There's now a toggle to ensure that PRs don't get built, so we no
longer have to check that manually. The 'branches' variable ensures
we only build for master.

The `--use-mirrors` option has also been deprecated for a while,
and now seems to be halting the build, so removed that also.