Implement transliteration of German umlauts
The previous handling of Unicode names just decomposed
the Unicode string to NFKD, which represents characters
with diacritic marks as a diacritic mark plus the
character itself. The following conversion to ASCII just
got rid of the diacritic marks.

This however, is not correct for German umlauts, which
should be transliterated to the character without the
diacritic mark followed by an "e".

The above code first converts the input to NFKC, correctly
replaces the German umlauts, and then converts to NFKD and
ASCII.

Unicode is hard.