Add build_and_deploy.sh
This script replaces .travis.yml. It is split into two stages.
The first stage serves as an outer wrapper responsible for
downloading the repository and executing the second-stage build
script inside a sandbox (this is the .travis.yml part).

In contrast to the old setup, the first stage is responsible for
uploading the built website to the webserver. This is much simpler
now since the build server will be inside the UW network.